refactor: complete modules
